RavenDB-22498 implemented /databases/*/indexes/auto/convert endpoint
